rename app to droneservice, modified .gitignore, move secrets to .gitignore'd file
