+++ backend/.gitignore
@@ -57,3 +57,6 @@ docs/_build/
 # PyBuilder
 target/
 
+# Application
+*.db
+secrets.py
