+++ backend/config.py
@@ -1,9 +1,11 @@
+import secrets
+
 # Statement for enabling the development environment
 DEBUG = True
 
 # Define the application directory
 import os
-BASE_DIR = os.path.abspath(os.path.dirname(__file__))  
+BASE_DIR = os.path.abspath(os.path.dirname(__file__))
 
 # Define the database - we are working with
 # SQLite for this example
@@ -20,8 +22,8 @@ THREADS_PER_PAGE = 2
 CSRF_ENABLED     = True
 
 # Use a secure, unique and absolutely secret key for
-# signing the data. 
-CSRF_SESSION_KEY = "secret"
+# signing the data.
+CSRF_SESSION_KEY = secrets.CSRF_SESSION_KEY
 
 # Secret key for signing cookies
-SECRET_KEY = "secret"
+SECRET_KEY = secrets.SECRET_KEY
