+++ backend/droneservice/__init__.py
@@ -1,0 +1,14 @@
+from flask import Flask, render_template
+from flask.ext.sqlalchemy import SQLAlchemy
+
+app = Flask(__name__)
+app.config.from_object('config')
+
+db = SQLAlchemy(app)
+db.create_all()
+
+
+
+@app.route('/')
+def hello_world():
+    return 'Hello World!'
