+++ backend/run.py
@@ -1,4 +1,3 @@
-# Run a test server.
-from app import app
+import droneservice
 
-app.run(host='0.0.0.0', port=8080, debug=True)
+droneservice.app.run(host='0.0.0.0', port=8000, debug=True)
